Refactor frontend continued
